AdVpnService: intent may be null in onStartCommand
As the Android documentation says:

"The Intent supplied to startService(Intent), as given. This may be null
if the service is being restarted after its process has gone away, and
it had previously returned anything except START_STICKY_COMPATIBILITY."

Assume that we want to (re-)start in that case.

This was discovered when stress testing with 5 different request
loops, 1 blocked, and 4 non-blocked hostnames.